refactor(server): extract session database metadata into server/sessionmeta
Carves the session-local catalog/metadata override helpers out of
server/ into a new server/sessionmeta subpackage with no dependency on
github.com/duckdb/duckdb-go.

The helpers install per-session DuckDB views and macros that make
current_database, pg_database, and information_schema_*_compat reflect
the client-visible database name on the PG wire — used by both the
standalone server's clientConn and the control plane's session bring-up.

Symbols moved (from server/session_database_metadata.go to
server/sessionmeta/sessionmeta.go):
  - InitSessionDatabaseMetadata (exported)
  - HasAttachedCatalog (was private hasAttachedCatalog; now exported)
  - All the buildSession* / sessionColumn* SQL builder helpers (private)
  - quoteSQLStringLiteral (private)

Type signature change: the QueryExecutor parameter is now
sqlcore.QueryExecutor (the interface moved to server/sqlcore in PR #488)
instead of server.QueryExecutor.

Backward compatibility preserved via re-export `var`s in
server/exports.go:

  var (
      HasAttachedCatalog          = sessionmeta.HasAttachedCatalog
      InitSessionDatabaseMetadata = sessionmeta.InitSessionDatabaseMetadata
  )

Internal callers updated:
  - server/conn.go now calls sessionmeta.InitSessionDatabaseMetadata and
    sessionmeta.HasAttachedCatalog directly
  - controlplane/control.go switches from server.X to sessionmeta.X
  - server/session_database_metadata{,_remote}_test.go keep their
    duckdb-go-using mocks but call sessionmeta.X for the moved symbols

  go list -deps ./server/sessionmeta | grep duckdb-go   # empty

Verified:
  - go build ./... clean
  - go build -tags kubernetes ./... clean
  - go test -short ./server/... — green except the pre-existing
    server/tlscert hang on real ACME calls (unrelated)
  - go list -deps ./server/sessionmeta | grep duckdb-go is empty
